requirements.txt
websockets
asyncio
protobuf
upstox-python-sdk
pandas
requests
pyotp
pytest-playwright
pyinstaller==4.5
numpy
